pyramid/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
google-generativeai==0.3.0
transformers==4.35.2
torch==2.1.1
python-multipart==0.0.6
